# Greatly speed up 'advanced' ipc receiving with big messages