--- a/lib/internal/child_process/serialization.js
+++ b/lib/internal/child_process/serialization.js
@@ -66,15 +66,11 @@ const advanced = {
     // Index 0 should always be present because we just pushed data into it.
     let messageBufferHead = channel[kMessageBuffer][0];
     while (messageBufferHead.length >= 4) {
-      // We read the uint manually here, because this is faster than first converting
+      // We call `readUInt32BE` manually here, because this is faster than first converting
       // it to a buffer and using `readUInt32BE` on that.
-      const size =
-        messageBufferHead[0] << 24 |
-        messageBufferHead[1] << 16 |
-        messageBufferHead[2] << 8 |
-        messageBufferHead[3];
+      const fullMessageSize = Buffer.prototype.readUInt32BE.call(messageBufferHead, 0) + 4;
 
-      if (channel[kMessageBufferSize] < 4 + size) break;
+      if (channel[kMessageBufferSize] < fullMessageSize) break;
 
       const concatenatedBuffer = channel[kMessageBuffer].length === 1 ?
         channel[kMessageBuffer][0] :
@@ -84,10 +80,10 @@ const advanced = {
         );
 
       const deserializer = new ChildProcessDeserializer(
-        TypedArrayPrototypeSubarray(concatenatedBuffer, 4, 4 + size)
+        TypedArrayPrototypeSubarray(concatenatedBuffer, 4, fullMessageSize)
       );
 
-      messageBufferHead = TypedArrayPrototypeSubarray(concatenatedBuffer, 4 + size);
+      messageBufferHead = TypedArrayPrototypeSubarray(concatenatedBuffer, fullMessageSize);
       channel[kMessageBufferSize] = messageBufferHead.length;
       channel[kMessageBuffer] =
         channel[kMessageBufferSize] !== 0 ? [messageBufferHead] : [];