--- a/lib/internal/child_process/serialization.js
+++ b/lib/internal/child_process/serialization.js
@@ -15,6 +15,7 @@ const assert = require('internal/assert');
 const { streamBaseState, kLastWriteWasAsync } = internalBinding('stream_wrap');
 
 const kMessageBuffer = Symbol('kMessageBuffer');
+const kMessageBufferSize = Symbol('kMessageBufferSize');
 const kJSONBuffer = Symbol('kJSONBuffer');
 const kStringDecoder = Symbol('kStringDecoder');
 
@@ -51,29 +52,51 @@ class ChildProcessDeserializer extends v8.DefaultDeserializer {
 //   (aka 'advanced')
 const advanced = {
   initMessageChannel(channel) {
-    channel[kMessageBuffer] = Buffer.alloc(0);
+    channel[kMessageBuffer] = [];
+    channel[kMessageBufferSize] = 0;
     channel.buffering = false;
   },
 
   *parseChannelMessages(channel, readData) {
     if (readData.length === 0) return;
 
-    let messageBuffer = Buffer.concat([channel[kMessageBuffer], readData]);
-    while (messageBuffer.length > 4) {
-      const size = messageBuffer.readUInt32BE();
-      if (messageBuffer.length < 4 + size) {
-        break;
-      }
+    channel[kMessageBuffer].push(readData);
+    channel[kMessageBufferSize] += readData.length;
+
+    // Index 0 should always be present because we just pushed data into it.
+    let messageBufferHead = channel[kMessageBuffer][0];
+    while (messageBufferHead.length >= 4) {
+      // We read the uint manually here, because this is faster than first converting
+      // it to a buffer and using `readUInt32BE` on that.
+      const size =
+        messageBufferHead[0] << 24 |
+        messageBufferHead[1] << 16 |
+        messageBufferHead[2] << 8 |
+        messageBufferHead[3];
+
+      if (channel[kMessageBufferSize] < 4 + size) break;
+
+      const concatenatedBuffer = channel[kMessageBuffer].length === 1 ?
+        channel[kMessageBuffer][0] :
+        Buffer.concat(
+          channel[kMessageBuffer],
+          channel[kMessageBufferSize]
+        );
 
       const deserializer = new ChildProcessDeserializer(
-        TypedArrayPrototypeSubarray(messageBuffer, 4, 4 + size));
-      messageBuffer = TypedArrayPrototypeSubarray(messageBuffer, 4 + size);
+        TypedArrayPrototypeSubarray(concatenatedBuffer, 4, 4 + size)
+      );
+
+      messageBufferHead = TypedArrayPrototypeSubarray(concatenatedBuffer, 4 + size);
+      channel[kMessageBufferSize] = messageBufferHead.length;
+      channel[kMessageBuffer] =
+        channel[kMessageBufferSize] !== 0 ? [messageBufferHead] : [];
 
       deserializer.readHeader();
       yield deserializer.readValue();
     }
-    channel[kMessageBuffer] = messageBuffer;
-    channel.buffering = messageBuffer.length > 0;
+
+    channel.buffering = channel[kMessageBufferSize] > 0;
   },
 
   writeChannelMessage(channel, req, message, handle) {