--- a/lib/internal/child_process/serialization.js
+++ b/lib/internal/child_process/serialization.js
@@ -100,20 +100,27 @@ const advanced = {
 
   writeChannelMessage(channel, req, message, handle) {
     const ser = new ChildProcessSerializer();
+    // Add 4 bytes, to later populate with message length
+    ser.writeRawBytes(Buffer.allocUnsafe(4));
     ser.writeHeader();
     ser.writeValue(message);
-    const serializedMessage = ser.releaseBuffer();
-    const sizeBuffer = Buffer.allocUnsafe(4);
-    sizeBuffer.writeUInt32BE(serializedMessage.length);
 
-    const buffer = Buffer.concat([
-      sizeBuffer,
-      serializedMessage,
-    ]);
-    const result = channel.writeBuffer(req, buffer, handle);
+    const serializedMessage = ser.releaseBuffer();
+    const serializedMessageLength = serializedMessage.length - 4;
+
+    serializedMessage.set([
+      serializedMessageLength >> 24 & 0xFF,
+      serializedMessageLength >> 16 & 0xFF,
+      serializedMessageLength >> 8 & 0xFF,
+      serializedMessageLength & 0xFF,
+    ], 0);
+
+    const result = channel.writeBuffer(req, serializedMessage, handle);
+
     // Mirror what stream_base_commons.js does for Buffer retention.
     if (streamBaseState[kLastWriteWasAsync])
-      req.buffer = buffer;
+      req.buffer = serializedMessage;
+
     return result;
   },
 };