--- a/lib/internal/child_process/serialization.js
+++ b/lib/internal/child_process/serialization.js
@@ -5,7 +5,6 @@ const {
   JSONStringify,
   StringPrototypeSplit,
   ArrayPrototypePush,
-  ReflectApply,
   Symbol,
   TypedArrayPrototypeSubarray,
 } = primordials;
@@ -15,7 +14,6 @@ const v8 = require('v8');
 const { isArrayBufferView } = require('internal/util/types');
 const assert = require('internal/assert');
 const { streamBaseState, kLastWriteWasAsync } = internalBinding('stream_wrap');
-const { readUInt32BE } = require('internal/buffer');
 
 const kMessageBuffer = Symbol('kMessageBuffer');
 const kMessageBufferSize = Symbol('kMessageBufferSize');
@@ -71,7 +69,12 @@ const advanced = {
     while (messageBufferHead.length >= 4) {
       // We call `readUInt32BE` manually here, because this is faster than first converting
       // it to a buffer and using `readUInt32BE` on that.
-      const fullMessageSize = ReflectApply(readUInt32BE, messageBufferHead, [0]) + 4;
+      const fullMessageSize = (
+        messageBufferHead[0] << 24 |
+        messageBufferHead[1] << 16 |
+        messageBufferHead[2] << 8 |
+        messageBufferHead[3]
+      ) + 4;
 
       if (channel[kMessageBufferSize] < fullMessageSize) break;