@rubikscraft 2022-05-01
@lpinca Is it normal for the osx tests to fail, or is that because this change?
Sorry, I'm not familiar with the node CI system.